pyproject.toml
[project]
name = "query2cmp"
version = "0.1.0"
description = "Natural language query → LLM → MCP invocation (testing & education)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.26.0",
    "openai>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
apple = [
    "apple-fm-sdk>=0.1.0",
]

[project.scripts]
query2cmp = "query2cmp.cli:main"

[tool.setuptools]
packages = ["query2cmp"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
